+++ business/database/migrations/v0.9.0.sql
@@ -126,3 +126,6 @@ ALTER TABLE `visited_tutorial_entries` ADD CONSTRAINT `fk_vts_user_id` FOREIGN K
 ALTER TABLE `user_storage` ADD `has_skipped_tutorial` BOOLEAN NOT NULL AFTER `max_energy`;
 
 ALTER TABLE `configuration` CHANGE `name` `name` VARCHAR(50) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL;
+
+-- v0.9.5
+ALTER TABLE `user_storage` ADD `can_alter_twitch_state` BOOLEAN NOT NULL AFTER `points`;