+++ game-frontend/projects/game-frontend/src/app/app.component.ts
@@ -6,6 +6,8 @@ import { state, style, trigger, transition, animate } from '@angular/animations'
 import { WebsocketService, UniverseGameService } from '@owge/universe';
 import { Player } from './types/twitch-player.type';
 import { TwitchState } from './types/twitch-state.type';
+import { TwitchService } from './services/twitch.service';
+import { delay, filter } from 'rxjs/operators';
 
 declare const Twitch: {
   Player: typeof Player
@@ -32,6 +34,7 @@ declare const Twitch: {
   styleUrls: ['./app.component.less', './app.component.scss']
 })
 export class AppComponent implements OnInit {
+  private static readonly _TWITCH_DIV = 'twitch-display';
 
   public isInGame: boolean;
 
@@ -56,7 +59,8 @@ export class AppComponent implements OnInit {
   public constructor(
     private _universeGameService: UniverseGameService,
     private _loadingService: LoadingService,
-    websocketService: WebsocketService
+    websocketService: WebsocketService,
+    twitchService: TwitchService
   ) {
     websocketService.isConnected.subscribe(val => {
       this.isConnected = val;
@@ -77,6 +81,13 @@ export class AppComponent implements OnInit {
       this.panicClass = panic ? 'panic-state' : '';
       this.isPanic = panic;
     });
+
+    twitchService.state().pipe(filter(twitchState => twitchState), delay(1000)).subscribe(() => {
+      if (this.hasToDisplayTwitch && this.hasToDisplayTwitch.hasToDisplay) {
+        this._removePlayer();
+        setTimeout(() => this._loadPlayer(this.hasToDisplayTwitch), 1000);
+      }
+    });
   }
 
   public ngOnInit() {
@@ -90,23 +101,29 @@ export class AppComponent implements OnInit {
 
   public onDisplayTwitch(val: TwitchState): void {
     this.hasToDisplayTwitch = val;
-    const divId = 'twitch-display';
     if (val.hasToDisplay) {
-      this.twitchPlayer = new Twitch.Player(divId, {
-        width: '100%',
-        height: '100%',
-        channel: 'kevinguanchedarias',
-        parent: [location.host.split(':')[0]]
-      });
-      if (!val.isPrimary) {
-        this.twitchPlayer.setMuted(true);
-      }
-      this.twitchPlayer.addEventListener(Twitch.Player.ONLINE, () => console.log('Channel is online'));
-      this.twitchPlayer.addEventListener(Twitch.Player.OFFLINE, () => console.log('Channel is offline'));
+      this._loadPlayer(val);
     } else {
-      this.twitchPlayer.pause();
-      delete this.twitchPlayer;
-      document.getElementById(divId).innerHTML = '';
+      this._removePlayer();
     }
+  }
+
+  private _loadPlayer(val: TwitchState) {
+    this.twitchPlayer = new Twitch.Player(AppComponent._TWITCH_DIV, {
+      width: '100%',
+      height: '100%',
+      channel: 'kevinguanchedarias',
+      parent: [location.host.split(':')[0]]
+    });
+    if (!val.isPrimary) {
+      this.twitchPlayer.setMuted(true);
+    }
+    this.twitchPlayer.addEventListener(Twitch.Player.ONLINE, () => console.log('Channel is online'));
+    this.twitchPlayer.addEventListener(Twitch.Player.OFFLINE, () => console.log('Channel is offline'));
+  }
+  private _removePlayer(): void {
+    this.twitchPlayer.pause();
+    delete this.twitchPlayer;
+    document.getElementById(AppComponent._TWITCH_DIV).innerHTML = '';
   }
 }