+++ game-frontend/projects/game-frontend/src/app/components/settings/settings.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
-import { WebsocketService } from '@owge/universe';
+import { User } from '@owge/core';
+import { UserStorage, WebsocketService } from '@owge/universe';
+import { TwitchService } from '../../services/twitch.service';
 
 @Component({
   selector: 'app-settings',
@@ -9,13 +11,20 @@ import { WebsocketService } from '@owge/universe';
 export class SettingsComponent {
 
   public isSyncing = false;
+  public user: User;
 
-  constructor(private _websocketService: WebsocketService) { }
+  constructor(private _websocketService: WebsocketService, private _twitchService: TwitchService, userStore: UserStorage<User>) {
+    userStore.currentUser.subscribe(user => this.user = user);
+  }
 
   public async triggerResync(): Promise<void> {
     this.isSyncing = true;
     await this._websocketService.clearCache();
     this.isSyncing = false;
+  }
+
+  public askTwitchState(): void {
+    this._twitchService.defineTwitchState(confirm('Are you live?')).subscribe();
   }
 
 }